Allow to disable all default mappings via g:no_CamelCaseMotion_maps
Here, it may make sense because some users may only want a subset of the mappings, and this saves them from defining a lot of dummy mappings. Also allow to do this selectively for either motions or text objects only.
Rename camelcasemotion#InnerMotion() to camelcasemotion#TextObject() and the corresponding s:Create...() function, too.